Refactor DES domain models ino des package
