+++ app/uk/gov/hmrc/individualsincomeapi/domain/JsonFormatters.scala
@@ -17,10 +17,13 @@
 package uk.gov.hmrc.individualsincomeapi.domain
 
 import java.util.UUID
+
 import play.api.libs.json._
+
 import scala.util.{Failure, Try}
 import play.api.libs.json.JodaWrites._
 import play.api.libs.json.JodaReads._
+import uk.gov.hmrc.individualsincomeapi.domain.des.{DesEmployment, DesEmploymentPayFrequency, DesEmployments, DesPayment}
 
 object JsonFormatters {
