+++ app/uk/gov/hmrc/individualsincomeapi/domain/SaSubmission.scala
@@ -20,6 +20,7 @@ import org.joda.time.LocalDate
 import play.api.libs.json.{Format, Json}
 import uk.gov.hmrc.domain.SaUtr
 import JsonFormatters._
+import uk.gov.hmrc.individualsincomeapi.domain.des.{DesAddress, DesSAIncome}
 
 case class SaFootprint(registrations: Seq[SaRegistration], taxReturns: Seq[SaTaxReturn])
