+++ app/uk/gov/hmrc/individualsincomeapi/domain/des/DesEmploymentIncome.scala
@@ -14,14 +14,15 @@
  * limitations under the License.
  */
 
-package uk.gov.hmrc.individualsincomeapi.domain
+package uk.gov.hmrc.individualsincomeapi.domain.des
 
 import org.joda.time.LocalDate
 import play.api.libs.functional.syntax._
-import play.api.libs.json.JodaReads._
-import play.api.libs.json.JodaWrites._
 import play.api.libs.json._
 import uk.gov.hmrc.domain.EmpRef
+import uk.gov.hmrc.individualsincomeapi.domain.Payment
+import play.api.libs.json.JodaWrites._
+import play.api.libs.json.JodaReads._
 
 case class DesEmployments(employments: Seq[DesEmployment])
