+++ app/uk/gov/hmrc/individualsincomeapi/services/v1/IncomeService.scala
@@ -24,7 +24,8 @@ import uk.gov.hmrc.http.{HeaderCarrier, Upstream5xxResponse}
 import uk.gov.hmrc.individualsincomeapi.connector.{DesConnector, IndividualsMatchingApiConnector}
 import uk.gov.hmrc.individualsincomeapi.domain.JsonFormatters._
 import uk.gov.hmrc.individualsincomeapi.domain.SandboxIncomeData.findByMatchId
-import uk.gov.hmrc.individualsincomeapi.domain.{DesEmployments, MatchNotFoundException, Payment}
+import uk.gov.hmrc.individualsincomeapi.domain.des.DesEmployments
+import uk.gov.hmrc.individualsincomeapi.domain.{MatchNotFoundException, Payment}
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
