+++ app/uk/gov/hmrc/individualsincomeapi/services/v1/SaIncomeService.scala
@@ -24,6 +24,7 @@ import uk.gov.hmrc.http.{HeaderCarrier, Upstream5xxResponse}
 import uk.gov.hmrc.individualsincomeapi.connector.{DesConnector, IndividualsMatchingApiConnector}
 import uk.gov.hmrc.individualsincomeapi.domain.SandboxIncomeData.findByMatchId
 import uk.gov.hmrc.individualsincomeapi.domain._
+import uk.gov.hmrc.individualsincomeapi.domain.des.DesSAIncome
 
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.Future
