+++ test/component/uk/gov/hmrc/individualsincomeapi/stubs/DesStub.scala
@@ -20,8 +20,9 @@ import com.github.tomakehurst.wiremock.client.WireMock._
 import play.api.http.Status
 import play.api.libs.json.Json
 import uk.gov.hmrc.domain.Nino
-import uk.gov.hmrc.individualsincomeapi.domain.{DesEmployments, DesSAIncome, TaxYear}
+import uk.gov.hmrc.individualsincomeapi.domain.TaxYear
 import uk.gov.hmrc.individualsincomeapi.domain.JsonFormatters._
+import uk.gov.hmrc.individualsincomeapi.domain.des.{DesEmployments, DesSAIncome}
 
 object DesStub extends MockHost(23000) {
