+++ test/unit/uk/gov/hmrc/individualsincomeapi/services/v1/IncomeServiceSpec.scala
@@ -31,6 +31,7 @@ import uk.gov.hmrc.http.{HeaderCarrier, Upstream5xxResponse}
 import uk.gov.hmrc.individualsincomeapi.connector.{DesConnector, IndividualsMatchingApiConnector}
 import uk.gov.hmrc.individualsincomeapi.domain.SandboxIncomeData._
 import uk.gov.hmrc.individualsincomeapi.domain._
+import uk.gov.hmrc.individualsincomeapi.domain.des.{DesEmployment, DesEmployments, DesPayment}
 import uk.gov.hmrc.individualsincomeapi.services.v1.{CacheId, LiveIncomeService, PayeIncomeCache, SandboxIncomeService}
 import unit.uk.gov.hmrc.individualsincomeapi.util.TestDates
 import utils.SpecBase
